agent_with_guardrails/requirements.txt
agno==2.5.3
google-generativeai==0.8.5
python-dotenv==1.0.1